feat(network): add host function

+++ include/network.h
@@ -32,7 +32,7 @@ public:
     int client();
 
 private:
-    int port = 6969;
+    unsigned short port = 6969;
 
     int recv_data(int &socket_fd, char *buffer, int size);
     int send_data(int &socket_fd, const char *buffer, int size);
